+++ usr/src/uts/common/sys/md5.h
@@ -60,7 +60,7 @@ typedef struct	{
 } MD5_CTX;
 
 void MD5Init(MD5_CTX *);
-void MD5Update(MD5_CTX *, const void *, unsigned int);
+void MD5Update(MD5_CTX *, const void *, size_t);
 void MD5Final(void *, MD5_CTX *);
 
 #ifdef	__cplusplus
